examples/registry/python/requirements.txt
minio>=7.2.0
boto3>=1.34.0
python-dotenv>=1.0.0
